+++ src/direction/hex_direction.rs
@@ -44,6 +44,10 @@ use crate::{DiagonalDirection, HexOrientation};
 #[repr(u8)]
 #[derive(Debug, Default, Copy, Clone, Eq, PartialEq, Hash)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[cfg_attr(
+    feature = "bevy_reflect",
+    derive(bevy_reflect::Reflect, bevy_reflect::FromReflect)
+)]
 pub enum Direction {
     #[default]
     /// Direction to (1, -1)