+++ src/lib.rs
@@ -32,6 +32,11 @@
 //!
 //! - `hexx = { version = "0.7", features = ["packed"] }`
 //!
+//! `hexx` supports [Bevy Reflection](https://docs.rs/bevy_reflect/latest/bevy_reflect) through the
+//! `bevy_reflect` feature. To enable it add the following line to your `Cargo.toml`:
+//!
+//! - `hexx = { version = "0.7", features = ["bevy_reflect"] }`
+//!
 //! ## Features
 //!
 //! `hexx` provides the [`Hex`] coordinates with: