+++ src/mesh/uv_mapping.rs
@@ -2,6 +2,10 @@ use glam::Vec2;
 
 #[derive(Debug, Clone)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[cfg_attr(
+    feature = "bevy_reflect",
+    derive(bevy_reflect::Reflect, bevy_reflect::FromReflect)
+)]
 /// Struct containing options for UV mapping.
 ///
 /// # Usage